requirements.txt
numpy>=1.22.0,<1.23.0
tensorboard==2.5.0
pandas==1.2.4
matplotlib==3.4.2
seaborn==0.11.1
pillow>=9.0.1
torchmeta>=1.8.0
tqdm==4.52.0
scipy==1.6.3
opencv-python==4.5.2.52
more-itertools==8.8.0


